server/pyproject.toml
[project]
name = "piscanner-server"
version = "0.1.0"
description = "Backend API to process PIScanner-client sent images"
authors = [
    {name = "email@nick-b.ca"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)"
]

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "mypy (>=1.18.2,<2.0.0)"
]

[tool.poetry]
package-mode = false
